+++ swanlab/data/run/callback/utils.py
@@ -1,0 +1,86 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+r"""
+@DATE: 2024/6/5 16:34
+@File: utils.py
+@IDE: pycharm
+@Description:
+    工具类
+"""
+from typing import Optional
+from swanlab.data.settings import SwanDataSettings
+from swanlab.log import swanlog
+from swanlab.utils.font import FONT
+from swanlab.env import is_windows
+from swanlab.package import get_package_version
+import os
+
+
+class U:
+    """
+    工具函数类，隔离SwanLabRunCallback回调与其他工具函数
+    """
+
+    def __init__(self):
+        self.settings: Optional[SwanDataSettings] = None
+
+    def inject(self, settings: SwanDataSettings):
+        """
+        为SwanLabRunCallback注入settings等一些依赖，因为实例化可能在SwanLabRun之前发生
+        :param settings: SwanDataSettings, 数据配置
+        :return:
+        """
+        self.settings = settings
+
+    @staticmethod
+    def formate_windows_path(path: str) -> str:
+        """这主要针对windows环境，输入的绝对路径可能不包含盘符，这里进行补充
+        主要是用于打印效果
+        如果不是windows环境，直接返回path，相当于没有调用这个函数
+
+        Parameters
+        ----------
+        path : str
+            待转换的路径
+
+        Returns
+        -------
+        str
+            增加了盘符的路径
+        """
+        if not is_windows():
+            return path
+        if not os.path.isabs(path):
+            return path
+        need_add = len(path) < 3 or path[1] != ":"
+        # 处理反斜杠, 保证路径的正确性
+        path = path.replace("/", "\\")
+        if need_add:
+            return os.path.join(os.getcwd()[:2], path)
+        return path
+
+    def _train_begin_print(self):
+        """
+        训练开始时的打印信息
+        """
+        swanlog.debug("SwanLab Runtime has initialized")
+        swanlog.debug("SwanLab will take over all the print information of the terminal from now on")
+        swanlog.info("Tracking run with swanlab version " + get_package_version())
+        local_path = FONT.magenta(FONT.bold(self.formate_windows_path(self.settings.run_dir)))
+        swanlog.info("Run data will be saved locally in " + local_path)
+
+    def _watch_tip_print(self):
+        """
+        watch命令提示打印
+        """
+        swanlog.info(
+            "🌟 Run `"
+            + FONT.bold("swanlab watch -l {}".format(self.formate_windows_path(self.settings.swanlog_dir)))
+            + "` to view SwanLab Experiment Dashboard locally"
+        )
+
+    def _train_finish_print(self):
+        """
+        打印结束信息
+        """
+        swanlog.info("Experiment {} has completed".format(FONT.yellow(self.settings.exp_name)))
